+++ lib/src/auth/config.dart
@@ -9,7 +9,26 @@ abstract class AuthorizationConfig {
 }
 
 abstract class AuthorizationStaus {
-  String platform;
+  final String platform;
+  final String username;
+  final String displayName;
+  final String identifier;
+  final String email;
+  final String phone;
+
+  final Map<String, dynamic> infos;
+  final Map<String, dynamic> beacons;
+
+  AuthorizationStaus({
+    this.platform,
+    this.username,
+    this.displayName,
+    this.identifier,
+    this.email,
+    this.phone,
+    this.infos,
+    this.beacons,
+  });
 
   Map<String, dynamic> toMap();
   String toString();