+++ lib/src/google/config.dart
@@ -1,5 +1,4 @@
 import 'package:authnull/src/auth/config.dart';
-import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 
 class GoogleConfig extends AuthorizationConfig {
@@ -14,21 +13,43 @@ class GoogleConfig extends AuthorizationConfig {
   });
 }
 
-class GoogleStatus extends AuthorizationStaus {
+class GoogleStatus implements AuthorizationStaus {
   @override
   final String platform = AuthorizationPlatform.Google;
 
-  final FirebaseUser user;
+  final String username;
+  final String displayName;
+  final String identifier;
+  final String email;
+  final String phone;
+
+  final Map<String, dynamic> infos;
+  final Map<String, dynamic> beacons;
+
+  final String token;
 
   GoogleStatus({
-    @required this.user,
+    @required this.username,
+    @required this.displayName,
+    @required this.identifier,
+    @required this.email,
+    @required this.phone,
+    @required this.infos,
+    @required this.beacons,
+    @required this.token,
   });
 
   @override
   Map<String, dynamic> toMap() {
     Map<String, dynamic> map = {
-      'displayName': this.user.displayName,
-      'email': this.user.email,
+      'username': this.username,
+      'displayName': this.displayName,
+      'identifier': this.identifier,
+      'email': this.email,
+      'phone': this.phone,
+      'infos': this.infos,
+      'beacons': this.beacons,
+      'token': this.token,
     };
     return map;
   }